Do not copy OSX nodes if OSX slave is not enabled

+++ docker/jenkins2-centos/contrib/jenkins/create-osx-node.sh
@@ -1,8 +1,12 @@
 #!/bin/sh
 
-echo "Copying Jenkins node configurations to ${JENKINS_HOME}/nodes ..."
-mkdir -p ${image_config_dir}
-cp -r ${image_config_dir}/nodes ${JENKINS_HOME}/nodes
-rm -rf ${image_config_dir}/nodes
+if [ ! -z "${OSX_SLAVE_HOST}" ]; then
+    echo "Copying Jenkins OSX node configuration to ${JENKINS_HOME}/nodes ..."
+    mkdir -p ${image_config_dir}
+    cp -r ${image_config_dir}/nodes ${JENKINS_HOME}/nodes
+    rm -rf ${image_config_dir}/nodes
 
-envsubst < "${JENKINS_HOME}/nodes/digger-ios/config.xml.tpl" > "${JENKINS_HOME}/nodes/digger-ios/config.xml"
+    envsubst < "${JENKINS_HOME}/nodes/digger-ios/config.xml.tpl" > "${JENKINS_HOME}/nodes/digger-ios/config.xml"
+else
+    echo "Skipping copying OSX node configuration since \${OSX_SLAVE_HOST} is not specified ..."
+fi
